Windows: faster rendering on 32-bit video cards. Audio fix for iOS.

+++ modules/juce_cryptography/encryption/juce_Primes.cpp
@@ -71,7 +71,7 @@ namespace PrimesHelpers
 
             i = (i - 1) >> 1;
 
-            while (i < numBits)
+            while (i < (unsigned int) numBits)
             {
                 result.setBit ((int) i);
                 i += prime;
